Upgrade to net8.0 using Upgrade Assistant

+++ DS2S META/Utils/ParamRows/ItemUsageRow.cs
@@ -20,7 +20,7 @@ namespace DS2S_META.Utils
             set
             {
                 _UsageBitfield = value;
-                WriteAtField(indUsageBitfield, BitConverter.GetBytes(value));
+                WriteAtField(indUsageBitfield, BitConverter.GetBytes((short)value));
             }
         }
